+++ rumqttc/src/eventloop.rs
@@ -43,7 +43,7 @@ pub enum ConnectionError {
     Tls(#[from] tls::Error),
     #[error("I/O: {0}")]
     Io(#[from] io::Error),
-    #[error("Connection refused, return code: {0:?}")]
+    #[error("Connection refused, return code: `{0:?}`")]
     ConnectionRefused(ConnectReturnCode),
     #[error("Expected ConnAck packet, received: {0:?}")]
     NotConnAck(Packet),