+++ rumqttc/src/v5/state.rs
@@ -2,7 +2,7 @@ use crate::v5::mqttbytes::PingReq;
 
 use super::mqttbytes::v5::Packet;
 use super::mqttbytes::{
-    self, ConnectReturnCode, PubAck, PubComp, PubRec, PubRel, Publish, QoS, Subscribe, Unsubscribe,
+    self, PubAck, PubComp, PubRec, PubRel, Publish, QoS, Subscribe, Unsubscribe,
 };
 use super::{Event, Incoming, Outgoing, Request};
 
@@ -16,9 +16,6 @@ pub enum StateError {
     /// Io Error while state is passed to network
     #[error("Io error: {0:?}")]
     Io(#[from] io::Error),
-    /// Broker's error reply to client's connect packet
-    #[error("Connect return code: `{0:?}`")]
-    Connect(ConnectReturnCode),
     /// Invalid state for a given operation
     #[error("Invalid state for a given operation")]
     InvalidState,